Assets updated to source folder (DISCARD CHANGES TO CLASSPATH ON PULL!)
When you launch the project from eclipse, it changes the classpath file and removes asset's change to a source folder. If you clone a fresh copy of the project, then you need to discard changes made to classpath, or change assets to be a source folder again. Otherwise, there will be a filenotfound exception.